requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pyserial==3.5
python-dotenv==1.0.1
Werkzeug==3.0.3
zipp==3.20.0
